build suggestion management
