--- app/tag/__init__.py
@@ -1,1 +1,0 @@
-from app.tag.views import tag  # noqa
