--- app/tag/forms.py
@@ -1,10 +1,0 @@
-from flask_wtf import Form
-from wtforms.fields import StringField, SubmitField, TextAreaField
-from wtforms.validators import InputRequired
-
-from app import db
-from app.models import Tag
-
-class TagForm(Form):
-    tag = StringField(validators=[InputRequired()])
-    submit = SubmitField()
