+++ app/templates/admin/index.html
@@ -32,7 +32,11 @@
                 {{ dashboard_option('Add New User', 'admin.new_user',
                                     description='Create a new user account', icon='add user icon') }}
                 {{ dashboard_option('Invite New User', 'admin.invite_user',
-                                    description='Invites a new user to create their own account', icon='add user icon') }}
+                                    description='Invite a new user to create their own account', icon='add user icon') }}
+                {{ dashboard_option('Manage Tags', 'admin.manage_tags',
+                                    description='Manage tags that can be applied to resources', icon='tag icon') }}
+                {{ dashboard_option('Review Suggestions', 'admin.review_suggestions',
+                                    description='Review suggested entries from users and contributors', icon='lightbulb icon') }}
             </div>
         </div>
     </div>
