+++ app/templates/admin/manage_user.html
