+++ app/templates/main/index.html
@@ -83,12 +83,6 @@
                       <div class="field">
                         <div class="ui checked checkbox">
                           <input type="checkbox" checked="">
-                          <label>Court Case</label>
-                        </div>
-                      </div>
-                      <div class="field">
-                        <div class="ui checked checkbox">
-                          <input type="checkbox" checked="">
                           <label>Other</label>
                         </div>
                       </div>
